+++ src/App.css
@@ -1,6 +1,29 @@
 :root {
+  /* width: 100vw;
+  height: 100vh; */
+  /* height: calc(var(--vh, 1vh) * 100); */
+}
+
+:root {
   font-family: "SUIT Variable", sans-serif;
   text-align: center;
-  width: 100vw;
-  height: calc(var(--vh, 1vh) * 100);
+  margin: 0 auto;
+  max-width: 430px;
+  width: 100%;
+  min-height: 100vh;
+  /* height: calc(var(--vh, 1vh) * 100); */
+  background-color: white;
+  box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
+  overflow-x: hidden;
+  overflow-y: hidden;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  /* justify-content: center; */
+  border-radius: 5px;
+  color: #1a1e27;
+  font-family: "SUIT Variable", -apple-system, BlinkMacSystemFont, Pretendard,
+    "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue",
+    sans-serif;
+  font-weight: 400;
 }
