+++ src/pages/Photobook/PhotoSelect.jsx
@@ -1,114 +1,79 @@
-import { useEffect, useState } from "react";
-import styles from "./Photoselect.module.scss";
-// import circleCheck from "../../img/circle-check-filled.png";
-import useAxios from "../../apis/axiosWithToken";
-import { useNavigate } from "react-router-dom";
+import React, { useState } from "react";
+import { useDrag } from "react-use-gesture";
+import { useSpring, animated } from "react-spring";
+import home from "../../img/home.png";
+import map from "../../img/map.png";
+import frame from "../../img/frame-line.png";
+import book from "../../img/book.png";
+import person from "../../img/person.png";
+import framelist from "../../img/framelist.png";
+import styles from "./PhotoSelect.module.scss";
+import Template from "../../component/MakeFrameCpn/Template";
+import FrameList from "../../component/AllFrameCpn/FrameList";
+import PhotoList from "./PhotoList";
+import { Link } from "react-router-dom";
 
 export default function PhotoSelect() {
-  const axios = useAxios();
-  const [getPhotos, setGetPhotos] = useState([]);
-  const [selectedPhotos, setSelectedPhotos] = useState([]);
-  const [isPhotobookExist, setIsPhotobookExist] = useState(false);
-  const router = useNavigate();
+  const [selectedButton, setSelectedButton] = useState("전체 포토");
+  // const [logoPos, setlogoPos] = useState({ x: 0, y: 0 });
+  const logoPos = useSpring({ x: 0, y: 0 });
+  const bindLogoPos = useDrag((params) => {
+    // setlogoPos({
+    //   x: params.offset[0],
+    //   y: params.offset[1],
+    // });
+    logoPos.x.set(params.offset[0]);
+    logoPos.y.set(params.offset[1]);
+  });
 
-  const handlePhotoClick = (index) => {
-    setSelectedPhotos((prevSelectedPhotos) => {
-      // 이미 선택된 사진이면 제거, 아니면 추가
-      if (prevSelectedPhotos.includes(index)) {
-        const newSelectedPhotos = prevSelectedPhotos.filter(
-          (selectedIndex) => selectedIndex !== index
-        );
-        console.log("삭제 후:", newSelectedPhotos);
-        return newSelectedPhotos;
-      } else {
-        const newSelectedPhotos = [...prevSelectedPhotos, index];
-        console.log("추가 후:", newSelectedPhotos);
-        return newSelectedPhotos;
-      }
-    });
+  const handleButtonClick = (button) => {
+    setSelectedButton(button);
   };
 
-  useEffect(() => {
-    console.log("세팅 후:", selectedPhotos);
-    console.log(
-      "api 보내기전: ",
-      selectedPhotos.map((index) => getPhotos[index])
-    );
-    axios
-      .get("/photo")
-      .then((response) => {
-        setGetPhotos(response.data.photoList);
-        setIsPhotobookExist(response.data.present);
-      })
-      .catch((error) => console.error(error));
-  }, [selectedPhotos]);
+  let middleContent;
 
-  const handlePostData = async () => {
-    const data = {
-      name: "토리",
-      addPhotoList: selectedPhotos.map((index) => getPhotos[index]),
-    };
-    
-    if (selectedPhotos.length === 0) {
-      // 선택된 사진이 없을 경우 알림 창 띄우기
-      alert("사진을 선택해주세요.");
-      return;
-    } 
-    else {
-      try {
-        const response = await axios.post("/photoBook", data);
-        console.log("API 응답:", response.data);
-        router("/photobook");
-      } catch (error) {
-        console.error("API 오류:", error);
-      }
-    }
-  };
-
-  const checkPhotobook = () => {
-    router("/photobook");
+  switch (selectedButton) {
+    case "프레임 제작":
+      middleContent = <FrameList />;
+      break;
+    case "포토북":
+      middleContent = <PhotoList />;
+      break;
+    default:
+      middleContent = <PhotoList />;
+      break;
   }
 
+  const topSpan = selectedButton === "프레임 제작" ? "Home" : "전체 포토";
+
   return (
-    <div className={styles.wrapper}>
-      <div className={styles.header}>
-        {isPhotobookExist ? (
-          <button onClick={checkPhotobook} className={styles.active}>
-            내 포토북 확인하기
+    <div className={styles.Frame}>
+      <div className={styles.Top}>
+        <Link to="/">
+          <button className={styles.ImgBtn}>
+            <img src={home} alt="logo" />
           </button>
-        ) : (
-          <button
-            onClick={handlePostData}
-            className={selectedPhotos.length !== 0 ? styles.active : styles.inactive}
-          >
-            나만의 포토북 제작하기
+        </Link>
+        <div className={styles.Center}>
+          <span>{topSpan}</span>
+        </div>
+        <div className={styles.Right}>
+          <button className={styles.mylistBtn}>
+            <img src={framelist} alt="framelist"></img>
           </button>
-        )}
+        </div>
       </div>
-      <div className={styles.bodySection}>
-        <div className={styles.boxWrapper}>
-          {/* 이미지 배열을 map 함수를 사용하여 렌더링 */}
-          {getPhotos.map((photo, index) => (
-            <div
-              key={index}
-              onClick={() => handlePhotoClick(index)}
-              className={styles.photoBox}
-            >
-              <img
-                className={styles.photo}
-                src={photo}
-                // width={120}
-                alt={`photo-${index}`}
-              />
-              {selectedPhotos.includes(index) && (
-                <div className={styles.selectBox}>
-                  <div className={styles.selectNumber}>
-                    {selectedPhotos.indexOf(index) + 1}
-                  </div>
-                </div>
-              )}
-            </div>
-          ))}
+      <div className={styles.Middle}>{middleContent}</div>
+      <div className={styles.Bottom}>
+        <div className={styles.ListBottom}>
+          <button onClick={() => handleButtonClick("프레임 제작")}>
+            <img src={frame} alt="frame"></img>
+            전체 프레임
+          </button>
+          <button onClick={() => handleButtonClick("포토북")}>
+            <img src={book} alt="book"></img>
+            포토북
+          </button>
         </div>
       </div>
     </div>
