+++ src/pages/Photobook/PhotoSelect.module.scss
@@ -1,87 +1,123 @@
-.wrapper {
-  width: 100vw;
-  height: 100vh;
-  overflow-x: hidden;
+.Frame {
+  width: 390px;
+  height: calc(var(--vh, 1vh) * 100);
+  background: #ffffff;
   display: flex;
   flex-direction: column;
-  gap: 30px;
+  justify-content: center;
+  user-select: none;
 }
 
-.header {
-  width: 100vw;
-  height: 12vh;
+.Top {
+  width: 100%;
+  height: 86px;
   display: flex;
+  flex-direction: row;
+  justify-content: space-between;
   align-items: center;
-  justify-content: center;
-  background-color: white;
-  margin-top: 20px;
-  margin-bottom: 20px;
-  .inactive, .active {
-    height: 50px;
-    width: 340px;
-    border-radius: 10px;
-    border: 2px solid #009eff;
-    color: #009eff;
-    text-align: center;
-    font-size: 14px;
-    font-weight: 500;
-    background-color: white;
-  }
-  .active {
-    background-color: #009eff;
-    color: white;
-  }
 }
 
-.bodySection {
-  width: 100vw;
-  min-height: 673px;
+.Top span {
+  flex: 1;
+  text-align: center;
+  font-size: 18px;
+  font-weight: 600;
+}
+
+.Top button {
+  width: 24px;
+  height: 24px;
+  background-color: #ffffff;
+  border: none;
+  padding-left: 30px;
+  padding-right: 50px;
+}
+
+.Center {
+  flex: 1; /* 가운데 영역이 남은 공간을 모두 차지하도록 설정 */
+  text-align: center; /* 텍스트 가운데 정렬 */
+}
+
+.mylistBtn {
+  width: 47px;
+  height: 47px;
+  font-size: 10px;
+  background-color: #ffffff;
+  border-radius: 30px;
+  margin-left: auto;
+  border: none;
+}
+
+.Middle {
+  width: 390px;
+  height: 689px;
   display: flex;
+  flex-direction: column;
   justify-content: center;
-  .boxWrapper {
-    position: relative;
-    width: 340px;
-    min-height: 673px;
-    display: grid;
-    grid-template-columns: repeat(2, 1fr); /* 2열로 나열 */
-    grid-template-rows: repeat(auto-fill, minmax(150px, 1fr));
-    grid-gap: 20px; /* 각 요소 사이의 간격 설정 */
-  }
-  .photoBox {
-    position: relative;
-    width: 160px;
-    height: 160px;
-    background-color: #f7f7f7;
-    img {
-      position: absolute;
-      z-index: 1;
-      height: 160px;
-      left: 26.22px;
-      width: auto;
-    }
-    .selectBox {
-      position: absolute;
-      z-index: 5;
-      width: 160px;
-      height: 160px;
-      background-color: rgba(255, 255, 255, 0.4);
-      border: 2.5px solid #009eff;
-      box-sizing: border-box;
-      display: flex;
-      flex-direction: row-reverse;
-      .selectNumber {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        width: 32px;
-        height: 32px;
-        color: white;
-        font-size: 14px;
-        font-weight: 600;
-        background-color: #009eff;
-        border-radius: 40px;
-        margin: 8px;
-      }
-    }
-  }
+  align-items: center;
+  background-color: #ffffff;
+  gap: 13px;
+}
+
+.Bottom {
+  width: 100%;
+  height: 70px;
+}
+
+.ListTop {
+  height: 40px;
+  padding: 14px 14px 14px 18px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  font-size: 10px;
+  font-weight: 400;
+}
+
+.ListTop button {
+  background-color: #ffffff;
+}
+
+.ListView {
+  display: inline-flex;
+  height: 160px;
+  padding: 15px 0px 15px 18px;
+  flex-direction: row;
+  gap: 18px;
+}
+
+.ListView span {
+  width: 72px;
+  height: 130px;
+  border-radius: 10px;
+  background: #d9d9d9;
+}
+
+.ListBottom {
+  width: 390px;
+  height: 70px;
+  background: #f7f8f8;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  align-items: center;
+  gap: 20px;
+}
+
+.ListBottom button {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  font-size: 10px;
+  font-weight: 400;
+  border: none;
+  background-color: #f7f8f8;
+}
+
+.ListBottom button img {
+  width: 24px;
+  height: 24px;
 }
